refactor(user): move change password mutation to hooks directory
The mutation was relocated from the api directory to the features/user/hooks/settings directory to better organize the codebase and follow the project's hook convention.
